+++ packages/wrangler/src/__tests__/cloudchamber/curl.test.ts
@@ -33,7 +33,7 @@ describe("cloudchamber curl", () => {
 		expect(helpStd.out).toMatchInlineSnapshot(`
 			"wrangler cloudchamber curl <path>
 
-			send a request to an arbitrary cloudchamber endpoint
+			send a request to an arbitrary Cloudchamber endpoint
 
 			POSITIONALS
 			  path  [string] [required] [default: \\"/\\"]
@@ -122,6 +122,7 @@ describe("cloudchamber curl", () => {
 			        \\"enabled\\": true
 			    },
 			    \\"network\\": {
+			        \\"mode\\": \\"public\\",
 			        \\"ipv4\\": \\"1.1.1.1\\"
 			    },
 			    \\"placements_ref\\": \\"http://ref\\",
@@ -171,9 +172,7 @@ describe("cloudchamber curl", () => {
 		);
 		expect(std.err).toMatchInlineSnapshot(`""`);
 		expect(std.out).toMatchInlineSnapshot(`
-			"├ Loading account
-			│
-			>> Body
+			">> Body
 			[
 			    {
 			        \\"id\\": \\"1\\",
@@ -190,6 +189,7 @@ describe("cloudchamber curl", () => {
 			            \\"enabled\\": true
 			        },
 			        \\"network\\": {
+			            \\"mode\\": \\"public\\",
 			            \\"ipv4\\": \\"1.1.1.1\\"
 			        },
 			        \\"placements_ref\\": \\"http://ref\\",
@@ -210,6 +210,7 @@ describe("cloudchamber curl", () => {
 			            \\"enabled\\": true
 			        },
 			        \\"network\\": {
+			            \\"mode\\": \\"public\\",
 			            \\"ipv4\\": \\"1.1.1.2\\"
 			        },
 			        \\"current_placement\\": {
@@ -240,6 +241,7 @@ describe("cloudchamber curl", () => {
 			            \\"enabled\\": true
 			        },
 			        \\"network\\": {
+			            \\"mode\\": \\"public\\",
 			            \\"ipv4\\": \\"1.1.1.1\\"
 			        },
 			        \\"placements_ref\\": \\"http://ref\\",
@@ -260,6 +262,7 @@ describe("cloudchamber curl", () => {
 			            \\"enabled\\": true
 			        },
 			        \\"network\\": {
+			            \\"mode\\": \\"public\\",
 			            \\"ipv4\\": \\"1.1.1.2\\"
 			        },
 			        \\"current_placement\\": {
@@ -332,8 +335,7 @@ describe("cloudchamber curl", () => {
 			"cloudchamber curl /deployments/v2 --header something:here"
 		);
 		expect(std.err).toMatchInlineSnapshot(`""`);
-		const text = std.out.split("\n").splice(2).join("\n");
-		const response = JSON.parse(text);
+		const response = JSON.parse(std.out);
 		expect(response.status).toEqual(500);
 		expect(response.statusText).toEqual("Unhandled Exception");
 	});