+++ packages/wrangler/src/api/startDevWorker/ConfigController.ts
@@ -1,8 +1,11 @@
 import assert from "node:assert";
 import path from "node:path";
+import { isDockerfile } from "@cloudflare/containers-shared";
 import { watch } from "chokidar";
 import { getAssetsOptions, validateAssetsArgsAndConfig } from "../../assets";
+import { fillOpenAPIConfiguration } from "../../cloudchamber/common";
 import { readConfig } from "../../config";
+import { containersScope } from "../../containers";
 import { getEntry } from "../../deployment-bundle/entry";
 import {
 	getBindings,
@@ -18,6 +21,7 @@ import {
 } from "../../environment-variables/misc-variables";
 import { UserError } from "../../errors";
 import { getFlag } from "../../experimental-flags";
+import { isNonInteractiveOrCI } from "../../is-interactive";
 import { logger, runWithLogLevel } from "../../logger";
 import { checkTypesDiff } from "../../type-generation/helpers";
 import {
@@ -52,7 +56,6 @@ import type {
 	StartDevWorkerOptions,
 	Trigger,
 } from "./types";
-import type { WorkerOptions } from "miniflare";
 
 type ConfigControllerEventMap = ControllerEventMap & {
 	configUpdate: [ConfigUpdateEvent];
@@ -158,7 +161,11 @@ async function resolveDevConfig(
 		enableContainers:
 			input.dev?.enableContainers ?? config.dev.enable_containers,
 		dockerPath: input.dev?.dockerPath ?? getDockerPath(),
-		containerEngine: input.dev?.containerEngine ?? getDockerHost(),
+		containerEngine:
+			input.dev?.containerEngine ??
+			config.dev.container_engine ??
+			getDockerHost(),
+		containerBuildId: input.dev?.containerBuildId,
 	} satisfies StartDevWorkerOptions["dev"];
 }
 
@@ -341,7 +348,7 @@ async function resolveConfig(
 			tsconfig: input.build?.tsconfig ?? config.tsconfig,
 			exports: entry.exports,
 		},
-		containers: resolveContainerConfig(config),
+		containers: config.containers,
 		dev: await resolveDevConfig(config, input),
 		legacy: {
 			site: legacySite,
@@ -400,6 +407,20 @@ async function resolveConfig(
 		);
 	}
 
+	// for pulling containers, we need to make sure the OpenAPI config for the
+	// container API client is properly set so that we can get the correct permissions
+	// from the cloudchamber API to pull from the repository.
+	const needsPulling = resolved.containers?.some(
+		(c) => !isDockerfile(c.image ?? c.configuration?.image)
+	);
+	if (needsPulling && !resolved.dev.remote) {
+		await fillOpenAPIConfiguration(
+			config,
+			isNonInteractiveOrCI(),
+			containersScope
+		);
+	}
+
 	// TODO(queues) support remote wrangler dev
 	const queues = extractBindingsOfType("queue", resolved.bindings);
 	if (
@@ -430,25 +451,6 @@ async function resolveConfig(
 	}
 
 	return resolved;
-}
-
-// TODO: move to containers-shared and use to merge config and args for container commands too
-function resolveContainerConfig(
-	config: Config
-): StartDevWorkerOptions["containers"] {
-	const containers: WorkerOptions["containers"] = {};
-	if (!config.dev.enable_containers) {
-		return containers;
-	}
-	for (const container of config.containers ?? []) {
-		containers[container.class_name] = {
-			image: container.image ?? container.configuration.image,
-			maxInstances: container.max_instances,
-			imageBuildContext: container.image_build_context,
-			name: container.name,
-		};
-	}
-	return containers;
 }
 
 export class ConfigController extends Controller<ConfigControllerEventMap> {