+++ packages/wrangler/src/cloudchamber/common.ts
@@ -1,37 +1,30 @@
-import { mkdir } from "fs/promises";
-import { logRaw, space, status, updateStatus } from "@cloudflare/cli";
+import { space, updateStatus } from "@cloudflare/cli";
 import { brandColor, dim } from "@cloudflare/cli/colors";
 import { inputPrompt, spinner } from "@cloudflare/cli/interactive";
 import {
 	ApiError,
 	DeploymentMutationError,
+	InstanceType,
 	OpenAPI,
 } from "@cloudflare/containers-shared";
-import { version as wranglerVersion } from "../../package.json";
+import {
+	addAuthorizationHeaderIfUnspecified,
+	addUserAgent,
+} from "../cfetch/internal";
 import { readConfig } from "../config";
-import { getConfigCache, purgeConfigCaches } from "../config-cache";
+import { constructStatusMessage } from "../core/CommandRegistry";
 import { getCloudflareApiBaseUrl } from "../environment-variables/misc-variables";
 import { UserError } from "../errors";
 import { isNonInteractiveOrCI } from "../is-interactive";
 import { logger } from "../logger";
-import {
-	DefaultScopeKeys,
-	getAccountFromCache,
-	getAccountId,
-	getAPIToken,
-	getAuthFromEnv,
-	getScopes,
-	logout,
-	reinitialiseAuthTokens,
-	requireAuth,
-	setLoginScopeKeys,
-} from "../user";
+import { requireApiToken, requireAuth } from "../user";
+import { printWranglerBanner } from "../wrangler-banner";
 import { parseByteSize } from "./../parse";
 import { wrap } from "./helpers/wrap";
 import { idToLocationName, loadAccount } from "./locations";
 import type { Config } from "../config";
 import type { CloudchamberConfig, ContainerApp } from "../config/environment";
-import type { Scope } from "../user";
+import type { containersScope } from "../containers";
 import type {
 	CommonYargsOptions,
 	StrictYargsOptionsToInterfaceJSON,
@@ -39,12 +32,23 @@ import type {
 import type { Arg } from "@cloudflare/cli/interactive";
 import type {
 	CompleteAccountCustomer,
+	CreateApplicationRequest,
 	EnvironmentVariable,
 	Label,
 	NetworkParameters,
+	UserDeploymentConfiguration,
 } from "@cloudflare/containers-shared";
 
+export const cloudchamberScope = "cloudchamber:write" as const;
+
 export type CommonCloudchamberConfiguration = { json: boolean };
+
+const containerIdRegexp = /[^/]{36}/;
+
+export function isValidContainerID(value: string): boolean {
+	const matches = value.match(containerIdRegexp);
+	return matches !== null;
+}
 
 /**
  * Regular expression for matching an image name.
@@ -103,7 +107,9 @@ export function handleFailure<
 		? K
 		: never,
 >(
-	cb: (args: CommandArgumentsObject, config: Config) => Promise<void>
+	command: string,
+	cb: (args: CommandArgumentsObject, config: Config) => Promise<void>,
+	scope: typeof cloudchamberScope | typeof containersScope
 ): (
 	args: CommonYargsOptions &
 		CommandArgumentsObject &
@@ -111,8 +117,15 @@ export function handleFailure<
 ) => Promise<void> {
 	return async (args) => {
 		try {
+			if (!args.json) {
+				await printWranglerBanner();
+				const commandStatus = command.includes("cloudchamber")
+					? "alpha"
+					: "open-beta";
+				logger.warn(constructStatusMessage(command, commandStatus));
+			}
 			const config = readConfig(args);
-			await fillOpenAPIConfiguration(config, args.json);
+			await fillOpenAPIConfiguration(config, args.json, scope);
 			await cb(args, config);
 		} catch (err) {
 			if (!args.json || !isNonInteractiveOrCI()) {
@@ -142,11 +155,16 @@ export async function loadAccountSpinner({ json }: { json?: boolean }) {
  * Gets the API URL depending if the user is using old/admin based authentication.
  *
  */
-async function getAPIUrl(config: Config) {
+async function getAPIUrl(
+	config: Config,
+	accountId: string,
+	scope: typeof cloudchamberScope | typeof containersScope
+) {
 	const api = getCloudflareApiBaseUrl(config);
-	// This one will probably be cache'd already so it won't ask for the accountId again
-	const accountId = config.account_id || (await getAccountId(config));
-	return `${api}/accounts/${accountId}/cloudchamber`;
+
+	const endpoint = scope === cloudchamberScope ? "cloudchamber" : "containers";
+
+	return `${api}/accounts/${accountId}/${endpoint}`;
 }
 
 export async function promiseSpinner<T>(
@@ -172,84 +190,22 @@ export async function promiseSpinner<T>(
 	return t;
 }
 
-export async function fillOpenAPIConfiguration(config: Config, json: boolean) {
+export async function fillOpenAPIConfiguration(
+	config: Config,
+	_json: boolean,
+	scope: typeof containersScope | typeof cloudchamberScope
+) {
 	const headers: Record<string, string> =
 		OpenAPI.HEADERS !== undefined ? { ...OpenAPI.HEADERS } : {};
 
-	// if the config cache folder doesn't exist, it means that there is not a node_modules folder in the tree
-	if (Object.keys(getConfigCache("wrangler-account.json")).length === 0) {
-		await wrap(mkdir("node_modules", {}));
-		purgeConfigCaches();
-	}
+	const accountId = await requireAuth(config);
+	const auth = requireApiToken();
+	addAuthorizationHeaderIfUnspecified(headers, auth);
+	addUserAgent(headers);
 
-	const scopes = getScopes();
-	const needsCloudchamberToken = !scopes?.find(
-		(scope) => scope === "cloudchamber:write"
-	);
-	const cloudchamberScope: Scope[] = ["cloudchamber:write"];
-	const scopesToSet: Scope[] =
-		scopes == undefined
-			? cloudchamberScope.concat(DefaultScopeKeys)
-			: cloudchamberScope.concat(scopes);
-
-	if (getAuthFromEnv() && needsCloudchamberToken) {
-		setLoginScopeKeys(scopesToSet);
-		// Wrangler will try to retrieve the oauth token and refresh it
-		// for its internal fetch call even if we have AuthFromEnv.
-		// Let's mock it
-		reinitialiseAuthTokens({
-			expiration_time: "2300-01-01:00:00:00+00:00",
-			oauth_token: "_",
-		});
-	} else {
-		if (needsCloudchamberToken && scopes) {
-			logRaw(
-				status.warning +
-					" We need to re-authenticate to add a cloudchamber token..."
-			);
-			// cache account id
-			await getAccountId(config);
-			const account = getAccountFromCache();
-			config.account_id = account?.id ?? config.account_id;
-			await promiseSpinner(logout(), { json, message: "Revoking token" });
-			purgeConfigCaches();
-			reinitialiseAuthTokens({});
-		}
-
-		setLoginScopeKeys(scopesToSet);
-
-		// Require either login, or environment variables being set to authenticate
-		//
-		// This will prompt the user for an accountId being chosen if they haven't configured the account id yet
-		const [, err] = await wrap(requireAuth(config));
-		if (err) {
-			throw new UserError(
-				`authenticating with the Cloudflare API: ${err.message}`
-			);
-		}
-	}
-
-	// Get the loaded API token
-	const token = getAPIToken();
-	if (!token) {
-		throw new UserError("unexpected apiToken not existing in credentials");
-	}
-
-	const val = "apiToken" in token ? token.apiToken : null;
-	// Don't try to support this method of authentication
-	if (!val) {
-		throw new UserError(
-			"we don't allow for authKey/email credentials, use `wrangler login` or CLOUDFLARE_API_TOKEN env variable to authenticate"
-		);
-	}
-
-	headers["Authorization"] = `Bearer ${val}`;
-	// These are being set by the internal fetch of wrangler, but we are not using it
-	// due to our OpenAPI codegenerated client.
-	headers["User-Agent"] = `wrangler/${wranglerVersion}`;
 	OpenAPI.CREDENTIALS = "omit";
 	if (OpenAPI.BASE.length === 0) {
-		const [base, errApiURL] = await wrap(getAPIUrl(config));
+		const [base, errApiURL] = await wrap(getAPIUrl(config, accountId, scope));
 		if (errApiURL) {
 			throw new UserError("getting the API url: " + errApiURL.message);
 		}
@@ -258,16 +214,6 @@ export async function fillOpenAPIConfiguration(config: Config, json: boolean) {
 	}
 
 	OpenAPI.HEADERS = headers;
-	const [, err] = await wrap(loadAccountSpinner({ json }));
-
-	if (err) {
-		let message = err.message;
-		if (json && err instanceof ApiError) {
-			message = JSON.stringify(err);
-		}
-
-		throw new UserError("Loading account failed: " + message);
-	}
 }
 
 export function interactWithUser(config: { json?: boolean }): boolean {
@@ -306,6 +252,7 @@ export function renderDeploymentConfiguration(
 	{
 		image,
 		location,
+		instanceType,
 		vcpu,
 		memoryMib,
 		environmentVariables,
@@ -315,6 +262,7 @@ export function renderDeploymentConfiguration(
 	}: {
 		image: string;
 		location: string;
+		instanceType?: InstanceType;
 		vcpu: number;
 		memoryMib: number;
 		environmentVariables: EnvironmentVariable[] | undefined;
@@ -352,13 +300,17 @@ export function renderDeploymentConfiguration(
 	const containerInformation = [
 		["image", image],
 		["location", idToLocationName(location)],
-		["vCPU", `${vcpu}`],
-		["memory", `${memoryMib} MiB`],
 		["environment variables", environmentVariablesText],
 		["labels", labelsText],
 		...(network === undefined
 			? []
 			: [["IPv4", network.assign_ipv4 === "predefined" ? "yes" : "no"]]),
+		...(instanceType === undefined
+			? [
+					["vCPU", `${vcpu}`],
+					["memory", `${memoryMib} MiB`],
+				]
+			: [["instance type", `${instanceType}`]]),
 	] as const;
 
 	updateStatus(
@@ -638,6 +590,38 @@ export async function promptForLabels(
 	return [];
 }
 
+export async function promptForInstanceType(
+	allowSkipping: boolean
+): Promise<InstanceType | undefined> {
+	let options = [
+		{ label: "dev: 1/16 vCPU, 256 MiB memory, 2 GB disk", value: "dev" },
+		{ label: "basic: 1/4 vCPU, 1 GiB memory, 4 GB disk", value: "basic" },
+		{ label: "standard: 1/2 vCPU, 4 GiB memory, 4 GB disk", value: "standard" },
+	];
+	if (allowSkipping) {
+		options = [{ label: "Do not set", value: "skip" }].concat(options);
+	}
+	const action = await inputPrompt({
+		question: "Which instance type should we use for your container?",
+		label: "",
+		defaultValue: false,
+		helpText: "",
+		type: "select",
+		options,
+	});
+
+	switch (action) {
+		case "dev":
+			return InstanceType.DEV;
+		case "basic":
+			return InstanceType.BASIC;
+		case "standard":
+			return InstanceType.STANDARD;
+		default:
+			return undefined;
+	}
+}
+
 // Return the amount of memory to use (in MiB) for a deployment given the
 // provided arguments and configuration.
 export function resolveMemory(
@@ -665,6 +649,96 @@ export function resolveAppDiskSize(
 	if (app === undefined) {
 		return undefined;
 	}
-	const disk = app.configuration.disk?.size ?? "2GB";
+	const disk = app.configuration?.disk?.size ?? "2GB";
 	return Math.round(parseByteSize(disk));
+}
+
+// Checks that instance type is one of 'dev', 'basic', or 'standard' and that it is not being set alongside memory or vcpu.
+// Returns the instance type to use if correctly set.
+export function checkInstanceType(
+	args: {
+		instanceType: string | undefined;
+		memory: string | undefined;
+		vcpu: number | undefined;
+	},
+	config: CloudchamberConfig
+): InstanceType | undefined {
+	const instance_type = args.instanceType ?? config.instance_type;
+	if (instance_type === undefined) {
+		return;
+	}
+
+	// If instance_type is specified as an argument, it will override any
+	// memory or vcpu specified in the config
+	if (args.memory !== undefined || args.vcpu !== undefined) {
+		throw new UserError(
+			`Field "instance_type" is mutually exclusive with "memory" and "vcpu". These fields cannot be set together.`
+		);
+	}
+
+	switch (instance_type) {
+		case "dev":
+			return InstanceType.DEV;
+		case "basic":
+			return InstanceType.BASIC;
+		case "standard":
+			return InstanceType.STANDARD;
+		default:
+			throw new UserError(
+				`"instance_type" field value is expected to be one of "dev", "basic", or "standard", but got "${instance_type}"`
+			);
+	}
+}
+
+// infers the instance type from a given configuration
+function inferInstanceType(
+	configuration: UserDeploymentConfiguration
+): InstanceType | undefined {
+	if (
+		configuration?.disk?.size_mb !== undefined &&
+		configuration?.memory_mib !== undefined &&
+		configuration?.vcpu !== undefined
+	) {
+		if (
+			configuration.disk.size_mb === 2000 &&
+			configuration.memory_mib === 256 &&
+			configuration.vcpu === 0.0625
+		) {
+			return InstanceType.DEV;
+		} else if (
+			configuration.disk.size_mb === 4000 &&
+			configuration.memory_mib === 1024 &&
+			configuration.vcpu === 0.25
+		) {
+			return InstanceType.BASIC;
+		} else if (
+			configuration.disk.size_mb === 4000 &&
+			configuration.memory_mib === 4096 &&
+			configuration.vcpu === 0.5
+		) {
+			return InstanceType.STANDARD;
+		}
+	}
+}
+
+// removes any disk, memory, or vcpu that have been set in an objects configuration. Used for rendering
+// diffs.
+export function cleanForInstanceType(
+	app: CreateApplicationRequest
+): ContainerApp {
+	if (!("configuration" in app)) {
+		return app as ContainerApp;
+	}
+
+	const instance_type = inferInstanceType(app.configuration);
+	if (instance_type !== undefined) {
+		app.configuration.instance_type = instance_type;
+	}
+
+	delete app.configuration.disk;
+	delete app.configuration.memory;
+	delete app.configuration.memory_mib;
+	delete app.configuration.vcpu;
+
+	return app as ContainerApp;
 }