+++ packages/wrangler/src/cloudchamber/images/list.ts
@@ -5,11 +5,13 @@ import {
 import { logger } from "../../logger";
 import { handleFailure, promiseSpinner } from "../common";
 import type { Config } from "../../config";
+import type { containersScope } from "../../containers";
 import type {
 	CommonYargsArgvJSON,
 	CommonYargsArgvSanitizedJSON,
 	StrictYargsOptionsToInterfaceJSON,
 } from "../../yargs-types";
+import type { cloudchamberScope } from "../common";
 import type { ImageRegistryPermissions } from "@cloudflare/containers-shared";
 
 interface CatalogResponse {
@@ -21,25 +23,36 @@ interface TagsResponse {
 	tags: string[];
 }
 
-export const imagesCommand = (yargs: CommonYargsArgvJSON) => {
+export const imagesCommand = (
+	yargs: CommonYargsArgvJSON,
+	scope: typeof containersScope | typeof cloudchamberScope
+) => {
 	return yargs
 		.command(
 			"list",
 			"perform operations on images in your Cloudflare managed registry",
 			(args) => listImagesYargs(args),
 			(args) =>
-				handleFailure(async (_args: CommonYargsArgvSanitizedJSON, config) => {
-					await handleListImagesCommand(args, config);
-				})(args)
+				handleFailure(
+					`wrangler containers images list`,
+					async (_args: CommonYargsArgvSanitizedJSON, config) => {
+						await handleListImagesCommand(args, config);
+					},
+					scope
+				)(args)
 		)
 		.command(
 			"delete [image]",
 			"remove an image from your Cloudflare managed registry",
 			(args) => deleteImageYargs(args),
 			(args) =>
-				handleFailure(async (_args: CommonYargsArgvSanitizedJSON, config) => {
-					await handleDeleteImageCommand(args, config);
-				})(args)
+				handleFailure(
+					`wrangler containers images delete`,
+					async (_args: CommonYargsArgvSanitizedJSON, config) => {
+						await handleDeleteImageCommand(args, config);
+					},
+					scope
+				)(args)
 		);
 };