+++ packages/wrangler/src/cloudchamber/ssh/ssh.ts
@@ -24,11 +24,13 @@ import {
 import { wrap } from "../helpers/wrap";
 import { validatePublicSSHKeyCLI, validateSSHKey } from "./validate";
 import type { Config } from "../../config";
+import type { containersScope } from "../../containers";
 import type {
 	CommonYargsArgvJSON,
 	CommonYargsArgvSanitizedJSON,
 	StrictYargsOptionsToInterfaceJSON,
 } from "../../yargs-types";
+import type { cloudchamberScope } from "../common";
 import type {
 	ListSSHPublicKeys,
 	SSHPublicKeyID,
@@ -106,23 +108,30 @@ export async function sshPrompts(
 	return key || undefined;
 }
 
-export const sshCommand = (yargs: CommonYargsArgvJSON) => {
+export const sshCommand = (
+	yargs: CommonYargsArgvJSON,
+	scope: typeof cloudchamberScope | typeof containersScope
+) => {
 	return yargs
 		.command(
 			"list",
 			"list the ssh keys added to your account",
 			(args) => args,
 			(args) =>
-				handleFailure(async (sshArgs: CommonYargsArgvSanitizedJSON, config) => {
-					// check we are in CI or if the user wants to just use JSON
-					if (!interactWithUser(sshArgs)) {
-						const sshKeys = await SshPublicKeysService.listSshPublicKeys();
-						console.log(JSON.stringify(sshKeys, null, 4));
-						return;
-					}
+				handleFailure(
+					`wrangler cloudchamber ssh list`,
+					async (sshArgs: CommonYargsArgvSanitizedJSON, config) => {
+						// check we are in CI or if the user wants to just use JSON
+						if (!interactWithUser(sshArgs)) {
+							const sshKeys = await SshPublicKeysService.listSshPublicKeys();
+							console.log(JSON.stringify(sshKeys, null, 4));
+							return;
+						}
 
-					await handleListSSHKeysCommand(sshArgs, config);
-				})(args)
+						await handleListSSHKeysCommand(sshArgs, config);
+					},
+					scope
+				)(args)
 		)
 		.command(
 			"create",
@@ -130,6 +139,7 @@ export const sshCommand = (yargs: CommonYargsArgvJSON) => {
 			(args) => createSSHPublicKeyOptionalYargs(args),
 			(args) =>
 				handleFailure(
+					`wrangler cloudchamber ssh create`,
 					async (
 						sshArgs: StrictYargsOptionsToInterfaceJSON<
 							typeof createSSHPublicKeyOptionalYargs
@@ -153,7 +163,8 @@ export const sshCommand = (yargs: CommonYargsArgvJSON) => {
 						}
 
 						await handleCreateSSHPublicKeyCommand(sshArgs);
-					}
+					},
+					scope
 				)(args)
 		);
 };